Additional tweaks to the GUI.
Change the "List DICOM Viewers" link in the tools menu to "DICOM Tools", and have it point to the wiki instead of a (hard coded) internal page.
Fix errors in the "Support" links, and increased the width of the pop-up window to better accomodate the rebranded wiki.
Change references to the development landing pages and wiki to the production instances.
Change the old support phone number to the new one, and mention the ability to text to the number.